Swap in Strict2 as default
