+++ bench/Main.hs
@@ -31,10 +31,10 @@ import qualified Data.Vector.Storable                         as DVS
 import qualified HaskellWorks.Data.FromForeignRegion          as IO
 import qualified HaskellWorks.Data.Sv.Internal.Char.Word64    as C
 import qualified HaskellWorks.Data.Sv.Lazy.Cursor             as SVL
-import qualified HaskellWorks.Data.Sv.Strict.Cursor           as SVS1
-import qualified HaskellWorks.Data.Sv.Strict.Cursor.Internal  as SVS1
-import qualified HaskellWorks.Data.Sv.Strict2.Cursor          as SVS2
-import qualified HaskellWorks.Data.Sv.Strict2.Cursor.Internal as SVS2
+import qualified HaskellWorks.Data.Sv.Strict.Cursor           as SVS2
+import qualified HaskellWorks.Data.Sv.Strict.Cursor.Internal  as SVS2
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor          as SVS1
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor.Internal as SVS1
 import qualified System.IO                                    as IO
 import qualified System.IO.MMap                               as IO
