+++ src/HaskellWorks/Data/Sv/Strict1/Cursor/Lens.hs
@@ -4,9 +4,9 @@
 {-# LANGUAGE TypeFamilies           #-}
 {-# LANGUAGE UndecidableInstances   #-}
 
-module HaskellWorks.Data.Sv.Strict2.Cursor.Lens where
+module HaskellWorks.Data.Sv.Strict1.Cursor.Lens where
 
 import Control.Lens
-import HaskellWorks.Data.Sv.Strict2.Cursor.Type
+import HaskellWorks.Data.Sv.Strict1.Cursor.Type
 
 makeFields ''SvCursor
