+++ test/HaskellWorks/Data/Sv/Strict1/Cursor/InternalSpec.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
-module HaskellWorks.Data.Sv.Strict2.Cursor.InternalSpec (spec) where
+module HaskellWorks.Data.Sv.Strict1.Cursor.InternalSpec (spec) where
 
 import Control.Concurrent
 import Control.Monad
@@ -29,10 +29,8 @@ import qualified Data.Vector.Storable                         as DVS
 import qualified HaskellWorks.Data.FromForeignRegion          as IO
 import qualified HaskellWorks.Data.Sv.Gen                     as G
 import qualified HaskellWorks.Data.Sv.Internal.Char.Word64    as C
-import qualified HaskellWorks.Data.Sv.Strict.Cursor           as SVS1
-import qualified HaskellWorks.Data.Sv.Strict.Cursor.Internal  as SVS1
-import qualified HaskellWorks.Data.Sv.Strict2.Cursor          as SVS2
-import qualified HaskellWorks.Data.Sv.Strict2.Cursor.Internal as SVS2
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor          as SVS
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor.Internal as SVS
 import qualified Hedgehog.Gen                                 as G
 import qualified Hedgehog.Range                               as R
 import qualified System.Directory                             as IO
@@ -42,43 +40,43 @@ import qualified System.Directory                             as IO
 {-# ANN module ("HLint: ignore Redundant bracket"   :: String) #-}
 
 spec :: Spec
-spec = describe "HaskellWorks.Data.Sv.Strict2.Cursor.InternalSpec" $ do
+spec = describe "HaskellWorks.Data.Sv.Strict1.Cursor.InternalSpec" $ do
   it "Case 1" $ requireTest $ do
     entries <- liftIO $ IO.listDirectory "data/bench"
     let files = ("data/bench/" ++) <$> (".csv" `isSuffixOf`) `filter` entries
     forM_ files $ \file -> do
       v <- liftIO $ IO.mmapFromForeignRegion file
-      let !pccccccc = DVS.foldr (\a b -> popCount1 a + b) 0 (SVS1.mkDsvRawBitsByWord64s      C.doubleQuote C.newline C.pipe      v)
-      let !actual   = DVS.foldr (\a b -> popCount1 a + b) 0 (fst $ SVS2.makeIndexes          '|'                                 v)
-      let !expected =     foldr (\a b -> popCount1 a + b) 0 (SVS1.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0  v)
+      let !pccccccc = DVS.foldr (\a b -> popCount1 a + b) 0 (SVS.mkDsvRawBitsByWord64s      C.doubleQuote C.newline C.pipe      v)
+      let !actual   = DVS.foldr (\a b -> popCount1 a + b) 0 (SVS.mkIbVector                 '|'                                 v)
+      let !expected =     foldr (\a b -> popCount1 a + b) 0 (SVS.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0  v)
       annotate $ "file    : " <> file
       annotate $ "pccccccc: " <> show pccccccc
       actual === expected
   it "Case 2" $ requireTest $ do
     bs :: ByteString <- forAll $ T.encodeUtf8 . T.pack <$> G.string (R.linear 0 128) (G.element " \"|\n")
     v <- forAll $ pure $ fromByteString bs
-    let !actual   = DVS.foldr (\a b -> popCount1 a + b) 0 (fst $ SVS2.makeIndexes          '|'                                 v)
-    let !expected =     foldr (\a b -> popCount1 a + b) 0 (SVS1.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0  v)
+    let !actual   = DVS.foldr (\a b -> popCount1 a + b) 0 (SVS.mkIbVector                 '|'                                 v)
+    let !expected =     foldr (\a b -> popCount1 a + b) 0 (SVS.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0  v)
     actual === expected
   it "Case 3" $ requireTest $ do
     bs :: ByteString <- forAll $ T.encodeUtf8 . T.pack <$> G.string (R.linear 0 128) (G.element " \"|\n")
     v <- forAll $ pure $ fromByteString bs
-    let !actual   = DVS.foldr (\a b -> popCount1 a + b) 0 (fst $ SVS2.makeIndexes          '|'                                 v)
-    let !expected =     foldr (\a b -> popCount1 a + b) 0 (SVS1.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0  v)
+    let !actual   = DVS.foldr (\a b -> popCount1 a + b) 0 (SVS.mkIbVector                 '|'                                 v)
+    let !expected =     foldr (\a b -> popCount1 a + b) 0 (SVS.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0  v)
     actual === expected
   it "Case 4" $ requireTest $ do
     bs :: ByteString <- forAll $ T.encodeUtf8 . T.pack <$> G.string (R.linear 0 10000) (G.element " \"")
     v <- forAll $ pure $ fromByteString bs
     numQuotes <- forAll $ pure $ BS.length $ BS.filter (== fromIntegral (ord '"')) bs
-    u <- forAll $ pure $ fst $ SVS2.makeIndexes '|' v
+    u <- forAll $ pure $ SVS.mkDsvRawBitsByWord64s C.doubleQuote C.newline C.pipe v
     let !pc = DVS.foldr (\a b -> popCount1 a + b) 0 u
     pc === fromIntegral numQuotes
   it "Case 5" $ requireTest $ do
     bs :: ByteString <- forAll $ T.encodeUtf8 . T.pack <$> G.string (R.linear 0 10000) (G.element " \"")
     v <- forAll $ pure $ fromByteString bs
     numQuotes <- forAll $ pure $ BS.length $ BS.filter (== fromIntegral (ord '"')) bs
-    u <- forAll $ pure $ fst $ SVS2.makeIndexes '|' v
-    let !expected = DVS.fromList $ SVS1.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0 v
+    u <- forAll $ pure $ SVS.mkDsvRawBitsByWord64s C.doubleQuote C.newline C.pipe v
+    let !expected = DVS.fromList $ SVS.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0 v
     let !pc = DVS.foldr (\a b -> popCount1 a + b) 0 u
     u === expected
   it "Case 6" $ requireTest $ do
@@ -86,7 +84,9 @@ spec = describe "HaskellWorks.Data.Sv.Strict2.Cursor.InternalSpec" $ do
     let files = ("data/bench/" ++) <$> (".csv" `isSuffixOf`) `filter` entries
     forM_ files $ \file -> do
       v <- liftIO $ IO.mmapFromForeignRegion file
-      let !actual   = fst $ SVS2.makeIndexes '|' v
-      let !expected = DVS.fromList $ SVS1.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0 v
+      let !actual   = SVS.mkIbVector '|' v
+      let !expected = DVS.fromList $ SVS.mkDsvInterestBitsByWord64s C.doubleQuote C.newline C.pipe 0 0 v
       annotate $ "file    : " <> file
       actual === expected
+
+
