+++ test/HaskellWorks/Data/SvSpec.hs
@@ -14,9 +14,9 @@ import HaskellWorks.Hspec.Hedgehog
 import Hedgehog
 import Test.Hspec
 
-import qualified Data.Vector.Storable                        as DVS
-import qualified HaskellWorks.Data.Sv.Strict.Cursor          as SVS
-import qualified HaskellWorks.Data.Sv.Strict.Cursor.Internal as SVS
+import qualified Data.Vector.Storable                         as DVS
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor          as SVS
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor.Internal as SVS
 
 {-# ANN module ("HLint: ignore Redundant do"        :: String) #-}
 {-# ANN module ("HLint: ignore Reduce duplication"  :: String) #-}
