+++ weigh/Space.hs
@@ -6,10 +6,10 @@ import Data.Vector                          (Vector)
 import HaskellWorks.Data.RankSelect.CsPoppy
 import Weigh
 
-import qualified Data.ByteString.Lazy               as LBS
-import qualified Data.Csv                           as CSV
-import qualified Data.Vector                        as DV
-import qualified HaskellWorks.Data.Sv.Strict.Cursor as SVS
+import qualified Data.ByteString.Lazy                as LBS
+import qualified Data.Csv                            as CSV
+import qualified Data.Vector                         as DV
+import qualified HaskellWorks.Data.Sv.Strict1.Cursor as SVS
 
 repeatedly :: (a -> Maybe a) -> a -> [a]
 repeatedly f a = a:case f a of
